theme update plus logic
