+++ src/app/components/teacherlist/teacherlist.component.html
@@ -13,7 +13,7 @@
             <div class="input-group">
 
                 <div class="input-group-append">
-                    <input type="text" class="form-control bg-light border-0 medium" placeholder="Search on Firstname" aria-label="Search" aria-describedby="basic-addon2" name="firstName" [(ngModel)]="firstName" (ngModelChange)="Search()">
+
                 </div>
             </div>
 
@@ -24,26 +24,34 @@
                     <thead>
 
                         <tr>
-                            <th (click)="Sort('employeeId')">Employee ID</th>
-                            <th (click)="Sort('firstName')">First Name</th>
-                            <th (click)="Sort('lastName')">Last Name</th>
-                            <th (click)="Sort('section')">Section</th>
+                            <th (click)="mysort('id')">Employee ID</th>
+                            <th (click)="mysort('first')">First Name</th>
+                            <th>Last Name</th>
+                            <th>Phone No</th>
+                            <th>Section</th>
+
+                            <th>Gender</th>
+                            <th>Salary</th>
                         </tr>
                     </thead>
 
                     <tbody>
-                        <tr *ngFor="let t of teachers | orderBy : key : reverse | paginate : {itemsPerPage:5,currentPage:p};">
+                        <tr *ngFor="let t of teachers">
                             <td>{{t.employeeId}}</td>
                             <td>{{t.firstName}}</td>
                             <td>{{t.lastName}}</td>
+                            <td>{{t.phone}}</td>
                             <td>{{t.section}}</td>
+
+                            <td>{{t.gender}}</td>
+                            <td>{{t.salary}}</td>
                         </tr>
 
 
                     </tbody>
 
                 </table>
-                <pagination-controls (pageChange)="p = $event"></pagination-controls>
+
             </div>
         </div>
     </div>
