+++ src/app/models/iteacher.ts
@@ -1,6 +1,9 @@
 export interface Iteacher {
-    employeeID: string;
+    employeeId: string;
     firstName: string;
     lastName: string;
+    phone: string;
     section: string;
+    gender: string;
+    salary: number;
 }
